Remove the interactivity fakeout
I don't know why it was there in the first place (well, because it
mirrors content-build), but we don't actually need it here. This
particular component will never render an interactive element (like a
link), so there's no reason to record a click event and act like it's
interactive by changing the cursor to a pointer.