WIP: Working on query unit tests (getting mock data)
Used the API explorer to get some mock data for the `vamcSystem` query module unit tests. This endpoint was failing due to circular dependencies when we include the `field_clinical_health_services` relation, and if that endpoint didn't fail, the frontend json viewer would. Added some extra logic in that endpoint handler to avoid that particular circular dependency