Move more lovell logic into the `facilitySideNav` function
instead of splitting it across multiple levels. It was originally added to the `lovel/staticProps` module in https://github.com/department-of-veterans-affairs/next-build/pull/226, but I don't think the original reason for spreading it out over multiple levels still holds. It's gotten more complicated, when it could be consolidated and simpler to reason about