Fix bad web component tag terminators
It didn't fix any errors, but I'm pretty sure web components can't have a self-closing tag, and it may not be correct even in React. We're currently treating these inconsistently throughout the app. Not sure if this actually needs to be fixed or not. I was seeing a funny error and was thinking that this might be a fix for it, but it wasn't.