Hardcode the "Mental health phone" label for facilities
The Drupal team disabled editing that field for the sake of consistency, so we often find that that label field is `null`. We instead have to define the lable in code for `health_care_local_facility`, `person_profile`, and `vamc_system_billing_insurance_form` nodes. In this case the `Phone` component is being used by health care local facilities