These props aren't being used, and they're duplicate information
Going to remove them for simplicity. We can always map this `vamcEhrSystem` value from those objects to the top-level prop if ever we encounter a use case where they would have been used. In this case I think the `regionPage` was the one we were "supposed" to use, but I ended up just giving it that vamcEhrSystem value directly. Or maybe it's not this regional page that was supposed to use it but instead the facility page that is already using it and drawing its `vamcEhrSystem` value from its region page relation