v3 grid: One way to define the columns
I don't like this because the width of this column changes sizes. Unfortunately the only way to do it the other way (see my code comments) is if we have a fixed width on one of the columns like existed before. We can't use the exact same rule from ye old `shame.css` because the width doesn't account for the column gaps, which are implemented with padding. The width of `.when-where-width` is too narrow. We could define our own new rule, but it seems a shame.

This wouldn't be a problem if these grids were using CSS grid.