Fix: Be more generic with the types
Those old types seem to be "pinned" to version 2 of the client options
(or some such), but that was causing a type error. Instead of using
those, we'll tie `createRedisClient` and `redisCache` together with a
more generic `ReturnType<typeof createClient>` to capitalize on the
default typing, which is what we were doing in the first place anyhow.

From what I can tell, our `redisCache` creates the abstraction layer
between `redis` and `next-drupal`, so as long as that API doesn't
change, this version bump should be otherwise transparent.