Move the older tests into that block too
I don't really think it's that valuable to test the formatter function in isolation when we can test the integrated form. We'd end up having to duplicate some tests anyway if we wanted to test all the same aspects of the data function too. And it's actually pretty easy to mock these drupal query functions anyway