Discovered that the map widget actually relied on global variables from the nearby VCs component!
In hte original liquid templates, the global variables are defined by `src/site/includes/vet_centers/nearby.liquid`, but they are not only needed by the widget element rendered in that template but also by the widget rendered by the `src/site/includes/image_and_static_map_dynamic.liquid` template! This relationship was not clear at all, and because I started with the satellite locations part of the template, I didn't have the global variables defined that it needed, and I was getting strange errors.