Ran the codemod
I selected all the codemods it gave me as options, the silly thing ran
for 1h 15m, and exited with a status code 1 after skipping the last
codemod of that list (next-request-geo-ip) because it's not deployed to
Vercel. I believe it then tried to run `yarn install` again, finished
that, then died because of a syntax error in the `yarn dlx
codemod@latest react/19/migration-recipe --no-interactive` command.