Path alias to the JS instead of the TS
This is because TS was importing the .ts file. The component library is
on an older version of React. React 19 (apparently) introduces stricter
types, so when we imported the `.ts` file, `tsc` would run the type
checks on the `.ts` file using the stricter types. This resulted in a
couple errors. By importing the `.js` file (in conjunction with our
`skipLibCheck` option set in `tsconfig.json`), we leave the type
checking to the library.

As far as I can tell, the stricter types have no effect at runtime.