Fix up a test
I believe this is due to a change in how React 19 handles boolean DOM
attribute values. I double-checked that just the presence of the
attribute had the same behavior, and from what I can tell, it does, so
changing the test to expect just the presence of the attribute (as an
empty string instead of `"true"`) is okay.