Add top CC content and troubleshoot testing issues with next-drupal
I was trying to use a the query formatter function in the new unit tests, but it was silently failing due to a check in the next-drupal-query package that would abort `queries.formatData` calls when `window` was defined, which it is for our Jest test environments by default for jsdom. Had to explicitly switch to a node environment for the query tests and had to stop trying to use the query formatter for mock data in the template tests.