Making the mocks a bit more accurate
I was thinking about making the `mock.formatted.ts` the expected value of a query formatter test, but it was getting very annoying to try to match up the big HTML string perfectly, even when I wrote the real output to a file and used that. Little things like that would be way harder to debug in a failing test than a snapshot test because there can be many mismatches for the single string.