Simplify the `LinkTeaser` component because of an a11y improvement
We don't want any of these links being rendered in a heading. They can just be bold-styled, regular links. We can also use the va-link for everything and use the `active` attribute instead of rendering our own icon for the (future) case of "spokes" links. (The spokes logic was added over 3 years ago, but it still isn't being used yet)