Give the LLMs more hints about what the `includes` parameter is for
I found that it was including fields in its calls that either weren't there at all or weren't necessary. Hopefully what I said has assured them that they probably don't need this in most cases. I tried it again, and it stopped trying to include fields that it didn't need to include.