Account for incomplete link teasers
The page /pittsburgh-health-care/research/safety-security had that problem. In prod the related links section isn't shown, so the content-build code must be ignoring the incomplete link and not showing the whole section. When you look at that page data, the related links section exists, but it just has a single link with no fields filled out.