Add the react widget to the page. Including TS garbage for history
I'll come back up and clean up my TypeScript garbage code. What I ended up committing was what I wanted to write in the first place in terms of type script utility types. This is the most correct. However, because of the exclusion/omission, it seems to give up on inferring the type of `fetched`. This isn't ideal, but it still works for now, and at least when someone reads the TypeScript, it will be logical. I wanted it to be able to be more strict about the properties of `fetched`. Whatever.