Make sure we use a root path that matches the current Lovell page
I spent a long time trying to understand why this content type is different from the others, and basically I gave up. I do know what's happening, though, and that it doesn't happen with any of the other bifurcated content types, and it doesn't happen with the non-bifurcated types (because there's no tricky path duplication business happening to those). What's happening for `/lovell-federal-health-care-tricare/work-with-us` as an example is that the entity path we're getting when we get to the query or formatter function is `...-va/work-with-us` instead of `...-tricare/work-with-us`.