Define web components (fix failed test)
The related test was failing due to the web components not being
rendered. The raw HTML caused axe to throw an accessibility violation,
even though it's actually fine.

This commit defines the VA web components for this test, and also
accesses href, text etc. as properties instead of attributes (since
that's how the related web components are defined)