Upgrade to latest version of the design system packages manually
My hunch is that removing the script include in the HEAD made automatic updating of the design system packages not work anymore. Related to that, my other hunch is that we were actually including the web components code twice! (once in our bundle and once as a direct include in the doc HEAD)