Fix the `AlertBlock` query and template
I was missing one of the nested includes for the case where an alert has block content. After that was fixed, I compared http://localhost:3999/louisville-health-care/health-services/suicide-prevention/ with production, and I found that it looked pretty messed up and was way different. Turns out that it didn't match the template in content-build, so I made it match, and now it looks better.