Refactor how we pass down and use heading levels and use `ParagraphList` recursively
This allows us to get the accordion headings correct in a deeply nested structure of QA paragraphs. Switched from using `headingLevel` to `currentHeadingLevel` so the paragraph itself could decide whether to pass that heading level on or increment it. That is, `headingLevel` was prescriptive from the parent, saying, "you should have this heading level", vs. with `currentHeadingLevel`, the parent is telling the paragraph, "this is the current heading level; do with that information what you will."

We do need to have a default `currentHeadingLevel` for any component that tries to increment it.