The `field_related_links` for these content types are `ListOfLinkTeaser` paragraphs, so use that
Made some simplifications and consolidated two queries into one. Now VAMC System and VAMC Facility use the same `ListOfLinkTeaser` formatter as other content types. One thing to note here is that the `componentParams` was used to set the `sectionHeader` to `''` by default anyway, and it wasn't until we got to the template rendering it that the template would set that value to something other than an empty string. That told me that there's no point in having it in the formatter, and we can keep the data simple and add it in the template when we're ready to render a whole list of link teasers.