Making a call and removing `name="map-directions"`
This is a deprecated property on anchor tags (https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/a#name) and doesn't get passed down by the web component to the anchor element (though it would be utterly useless anyway inside the shadow DOM). The only argument I could make for keeping it around is that there could be some script somewhere that will focus on that link using the name or an internal page anchor. I think that's highly unlikely, and it would only be useful on pages that only have one (not the listing pages).