Don't show link summaries for VAMC System and Facility pages
I found that the link data for [/boston-health-care/locations/brockton-va-medical-center](https://www.va.gov/boston-health-care/locations/brockton-va-medical-center/) actually includes a link summary for one of the links that we were previously hiding via the data formatter. I took a look at both the old data formatters and realized that both pages were doing that, so I changed the `RelatedLinks` React component used by those two pages so it doesn't show the summary, since it seems to be more of a display decision.