Change back to use the VAMC system title
When doing some manual VRT, I noticed the 'Other services at' heading
was different from prod. This _was_ a merge conflict, and I'm not sure
what the story is, but my change here reverts it back to what prod
shows.

I wish Laura were here for me to ask which was correct. 😢