Temporarily committing some data that I collected on breadcrumbs across content types
Pulled down 4 entities from every page content type and recorded the last breadcrumb to see how that data varied from content type to content type. I found that about half the time the final breadcrumb's URI was `internal:#` and half it was the full entity path. I've chosen to normalize this in my function that does the default processing of breadcrumbs.