Use the `crop_square` size like the original
I thought I could choose a different one that was represented already in the json, but it turns out that the one I chose wasn't being generated by Drupal, even though the json API returned an entry for it. That is, the image file wasn't being generated. What I'll need to do instead is make a change in Drupal to include the `crop_square` derivative in the json for all images...for ALL images.