News spotlight section: Can achieve the full-width color split with custom css
We're not using the old classes anymore because they don't match our new breakpoints.

This also seems like a more straightforward solution to me than using a background gradient...though now that I think about it, I end up throwing out the custom styles on the smaller screens anyway and apply the background color directly to the content elements, which is the part that I wanted to be more straightforward anyway. I might follow up with another commit to further simplify this